backend-python/requirements.txt
grpcio==1.76.0
protobuf==6.33.2
google-cloud-spanner==3.49.1
